Experimental GPU performance optimizations
This is a highly speculative patch. DO NOT SHIP IT INTO PRODUCTION!

The goal is to reduce the amount of work inflicted on the GPU,
by extending delta-frame damage tracking all the way down to
the compute and fragment shaders.

For each frame to render on the GPU, the damage is represented as a
Rect of cells covering the rectangular area that must be redrawn.
Without this patch, this area is always equal to all cells, even if
the copying of cells from CPU to GPU is already (partially) optimized
by Frame::deltaCopyCells ().

The area to be rendered is represented by the newly added renderExtent
of CharVdev. Then, dispatchCompute is invoked with only the cell group
needed. The draw program is also restricted by reconfiguring the
render quad to only cover the affected damage rect.

I am convinced that the code does what I intend it to, but somewhat
paradoxically, I still have a hard time seeing any kind of real
performance improvement from this. It might be that the load
represented by Zutty on a GPU is so little that it simply does not
matter. Still keeping this code around for possible later reference.