Fix signedness bug in Vterm::insertRows()
Countdown loop with unsigned loop counter: FAIL.

This made Zutty crash on receiving an 'Insert Line' sequence with
cursor on the top row and a non-default argument greater than the
number of terminal rows (CSI <n> L, where n > rows). For example,
with the default geometry of 24 rows:

    printf "\e[1;1H\e[48L"

Many thanks to Nathan Poysti for finding this bug, narrowing it down
and reporting.